requirements-client.txt
python-socketio
argparse
websocket-client